% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_June2021.R
\name{age.key}
\alias{age.key}
\title{Create age-length key}
\usage{
age.key(dataset, age_column, length_column, length_unit = NA)
}
\arguments{
\item{dataset}{dataset containing at least 2 columns, one containing an age field, and one containing length field}

\item{age_column}{string specifying the name of the column containing ages (in years)}

\item{length_column}{string specifying the name of the column containing lengths}

\item{length_unit}{string specifying the length units options are: "mm" = millimeters; "cm" = centimeters, if none specified defaults to "mm"}
}
\value{
returns age-length key
}
\description{
Generate age-length key from a subsample of aged fish
}
\examples{
data(loch_blcr_2020_age)
loch_blcr_alk <- age.key(dataset=loch_blcr_2020_age, age_column="age", length_column="tl", length_unit = 'mm')
}
