% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_June2021.R
\name{cpue.plot}
\alias{cpue.plot}
\title{Catch-Per-Unit-Effort Timeseries Plot}
\usage{
cpue.plot(
  datafile,
  speciesList = list(),
  species_size_strata = list(),
  years = list(),
  seasons = list(),
  print = FALSE,
  figure_filename = NA,
  fig_scale = 1
)
}
\arguments{
\item{datafile}{data, should be output from ltm.data.summary function}

\item{speciesList}{list of selected species, can specify by common name, scientific name or species code}

\item{species_size_strata}{optional argument, specifies size strata groups for which CPUEs should be calculated, see example below for proper convention.}

\item{years}{list of years to include in figure}

\item{seasons}{---currently functionless, will update in future version, to avoid errors make sure that all seasons in input dataset are the same---}

\item{print}{boolean, if TRUE figure will be saved to file}

\item{figure_filename}{if print=TRUE, figure will be saved to this filename}

\item{fig_scale}{adjust to scale output figure size}
}
\value{
Returns the summarized data used to construct the plot
}
\description{
Create CPUE timeseries for species and years selected
}
\examples{
# import and format data
data(newnans)
newn_sum <- ltm.data.summary("Newnans Lake", newnans)
# CPUE plots for bluegill, largemouth bass, and brown bullheads
newnans_cpue <- cpue.plot(newn_sum,
 speciesList=c("BLUE","LMB", "BRBU"),
 years = c(2016:2020))
# CPUE plots by size class
newnans_cpue2 <- cpue.plot(newn_sum,
 speciesList=c("BLUE","LMB","BRBU"),
 species_size_strata = list(
  BLUE = list(
   YOY = c(0,8),
   Quality = c(18,50)),
  LMB = list(
   YOY = c(0,20),
   Quality = c(30,50),
   Trophy = c(51,100))
 ),
 years = c(2016:2020)
)
}
