% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ltm_report.R
\name{create_ltm_report}
\alias{create_ltm_report}
\title{Create a new LTM report template}
\usage{
create_ltm_report(dir_name, report_name = dir_name, dir_root = getwd())
}
\arguments{
\item{dir_name}{*string* specifying the name of the report directory to create}

\item{report_name}{*string* specifying the report name, used in file names so avoid spaces
defaults to name specified in directory}

\item{dir_root}{*string* specifying the path to the root directory in which to create the report project. 
Defaults to the current working directory.}
}
\value{
boolean - returns `TRUE` if the function executes without error
}
\description{
Create a new LTM report template
}
\examples{

\dontrun{
create_ltm_report(dir_name = "test_report", 
                  report_name ="LakeDoe_LMB_2020")
}

}
