% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{deprecate}
\alias{deprecate}
\title{INTERNAL: Create a deprecation warning}
\usage{
deprecate(new_function)
}
\value{
warning message
}
\description{
INTERNAL: Create a deprecation warning
}
\examples{
deprecated_function <- function() {
  deprecate("new_function")
}
deprecated_function()
}
