% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_table}
\alias{fish_table}
\title{LTM Fish Lookup Table}
\format{
A dataframe with 321 obs and 7 variables
\describe{
\item{F_code}{Species field code}
\item{common_name}{common name}
\item{scientific_name}{binomial scientific name}
\item{eco_type}{categorical variable grouping species as "Exotic", "Fresh", or "Salt"}
\item{family_scientific}{family name, scientific (latin)}
\item{family_common}{family name, common}
\item{fwc_id}{unique identification number for species, can be used as primary key}}
}
\usage{
data(fish_table)
}
\description{
Species lookup table for freshwater fish.
}
\keyword{datasets}
