% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flfishltm.R
\docType{package}
\name{flfishltm-package}
\alias{flfishltm}
\alias{flfishltm-package}
\title{flfishltm: A package for summarizing and analyzing FL FWC LTM Data}
\description{
ADD PACKAGE DETAILS
See
}
\details{
\code{vignette("flfishltm", package="flfishltm")}

intro to flfishltm for more details
}
\section{flfishltm functions}{


\code{\link{ltm.data.summary}}

\code{\link{species.history}}

\code{\link{len.dist}}
}

\author{
\strong{Maintainer}: Jason O'Connor \email{jason.oconnor@myfwc.com}

Authors:
\itemize{
  \item Jason O'Connor \email{jason.oconnor@myfwc.com}
}

}
