% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpt.R
\name{get_wpt}
\alias{get_wpt}
\title{Extract waypoint data from .gpx files}
\usage{
get_wpt(gpx)
}
\arguments{
\item{gpx}{A .gpx file generated using a Garmin GPSmap 78s}
}
\value{
data frame containing point ID, lat, long and elevation
}
\description{
Extract waypoint data from .gpx files
}
\examples{
# df <- get_wpt('Waypoints_01-DEC-22.gpx')
}
