% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_June2021.R
\name{guild.comp}
\alias{guild.comp}
\title{Guild Composition Plot}
\usage{
guild.comp(
  ltm_dataset,
  waterbody_list = list(),
  years = list(),
  color_scale = "default",
  save = FALSE
)
}
\arguments{
\item{ltm_dataset}{data, should be output from ltm.data.summary function}

\item{waterbody_list}{---currently functionless, will update in future version---}

\item{years}{list of years to include in figure}

\item{color_scale}{color scale for output figure defaults to default r color scale, other options include "grayscale" for black and white, and "viridis" for viridis magma scale}

\item{save}{boolean, if TRUE figure will be saved to file}
}
\value{
returns summarize data used to construct figures
}
\description{
Create Guild Composition plot
}
\examples{
data(newnans)
newn_sum <- ltm.data.summary(file=newnans)
newnans_guild <- guild.comp(newn_sum)
}
