% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_addgapyears}
\alias{helper_addgapyears}
\title{Helper function to add blank rows to account for missing years in a yearly summary table}
\usage{
helper_addgapyears(sumTable)
}
\arguments{
\item{sumTable}{\code{data.frame} or \code{tibble}}
}
\value{
table
}
\description{
currently used within \code{\link[=ltm.data.summary]{ltm.data.summary()}}
}
\examples{

test_data <- structure(list(
Year = c("2006 - 2007", "2007 - 2008", "2008 - 2009","2009 - 2010", 
          "2010 - 2011", "2012 - 2013", "2013 - 2014", "2014 - 2015", 
          "2015 - 2016", "2019 - 2020", "2020 - 2021"),
yr = c(2006, 2007, 2008, 2009, 2010, 2012,2013, 2014, 2015, 2019, 2020),
BOW = c(0.226666666666667, 0.24, 0.188, 0.252, 0.132, 0.02, 0.116, 0.112, 
        0.064, 0.26, 0.128), LMB = c(0.00666666666666667, 0.0733333333333333, 
        0.064, 0.084, 0.04, 0.02, 0.048, 0.148, 0.212, 0.136, 0.12)), 
        row.names = c(NA, -11L), class = c("tbl_df", "tbl", "data.frame"))
        
helper_addgapyears(test_data) 
 

}
