% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_June2021.R
\name{is.outlier}
\alias{is.outlier}
\title{Check if Outlier}
\usage{
is.outlier(x)
}
\arguments{
\item{x}{vector of numerical values}
}
\value{
boolean
}
\description{
Checks whether value is an outlier
}
\examples{
numbers = c(1:10,1:10,1:10,1000)
numbers_outlier <- is.outlier(numbers)
numbers[numbers_outlier]#'
}
