% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_Import.R
\name{is.StdFishQry}
\alias{is.StdFishQry}
\title{Standard Fish Query Test}
\usage{
is.StdFishQry(x)
}
\arguments{
\item{x}{dataset}
}
\value{
true/false
}
\description{
Check if data is from LTM Standard Fish Query
}
\examples{
data(newnans)
is.StdFishQry(newnans)
####
is.StdFishQry(newnans[, -1])
###
new2 <- newnans
new2$newcolumn <- NA
is.StdFishQry(new2)
}
