% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_June2021.R
\name{len.dist}
\alias{len.dist}
\title{Length Distribution Plot}
\usage{
len.dist(
  datafile,
  speciesList = list(),
  years = list(),
  seasons = list(),
  print = FALSE,
  figure_filename = NA,
  fig_scale = 1
)
}
\arguments{
\item{datafile}{data, should be output from ltm.data.summary function}

\item{speciesList}{list of selected species, can specify by common name, scientific name or species code}

\item{years}{list of years to include in figure}

\item{seasons}{---currently functionless, will update in future version, to avoid errors make sure that all seasons in input dataset are the same---}

\item{print}{boolean, if TRUE figure will be saved to file}

\item{figure_filename}{if print=TRUE, figure will be saved to this filename}

\item{fig_scale}{adjust to scale output figure size}
}
\value{
Returns the summarized data used to construct the histogram plot
}
\description{
Create Length Distribution histograms for species and years selected
}
\examples{
data(newnans)
newn_sum <- ltm.data.summary("Newnans",newnans)
newnans_LD <- len.dist(newn_sum, speciesList=c("BLUE","LMB", "BRBU"), years = c(2016:2020))
}
