% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len_dist.R
\name{len_dist}
\alias{len_dist}
\title{Length Distribution Plot}
\usage{
len_dist(
  datafile,
  speciesList = list(),
  years = list(),
  seasons = list(),
  print = FALSE,
  figure_filename = NA,
  fig_scale = 1,
  return_object = "data"
)
}
\arguments{
\item{datafile}{data, should be output from ltm.data.summary function}

\item{speciesList}{list of selected species, can specify by common name, scientific name or species code}

\item{years}{list of years to include in figure}

\item{seasons}{---currently functionless, will update in future version, to avoid errors make sure that all seasons in input dataset are the same---}

\item{print}{boolean, if TRUE figure will be saved to file}

\item{figure_filename}{if print=TRUE, figure will be saved to this filename}

\item{fig_scale}{adjust to scale output figure size}

\item{return_object}{\emph{string} if "data" then function returns the dataset
that is used to generate the figure, else if "ggplot" then function returns
ggplot object}
}
\value{
either summarised data or a ggplot object. Specify the return object type with \code{return_object}
}
\description{
Create Length Distribution histograms for species and years selected
}
\examples{
data(newnans)
newn_sum <- ltm.data.summary("Newnans",newnans)
newnans_LD <- len_dist(newn_sum, speciesList=c("BLUE","LMB", "BRBU"), years = c(2016:2020))
}
