% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_June2021.R
\name{ltm.data.summary}
\alias{ltm.data.summary}
\title{LTM Data Summary}
\usage{
ltm.data.summary(
  waterbodyname = "No Waterbody Specified",
  file,
  outtables = 0,
  printfigs = 0,
  print_directory = getwd()
)
}
\arguments{
\item{waterbodyname}{The name of the selected Waterbody as you would like it to appear in filenames and figure titles}

\item{file}{path to the raw query file, should be a .csv file}

\item{outtables}{option to specify which summary tables to save to working directory, default 0 exports no summary tables}

\item{printfigs}{option to specify which figures to save to print directory, default 0 prints no figures, 1 prints all figures}

\item{print_directory}{specify directory to save exported figures}
}
\value{
list of summarized datasets
}
\description{
Summarize raw LTM data from "Standard Fish Query"
}
\examples{
#Import demo data for Newnans Lake
data(newnans)
#dplyr::summarise Newnans Lake data
newn_sum <- ltm.data.summary("Newnans Lake", file = newnans)
# Access summary tables within the newn_sum object
newn_sum$SpeciesList
newn_sum$CPUE_number
newn_sum$RawData
newn_sum$Comp_num
}
