% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ltm.import}
\alias{ltm.import}
\title{Import data from FWC LTM Query}
\usage{
ltm.import(datafile)
}
\arguments{
\item{datafile}{an ltm query dataset, can be either ".csv" or ".Rdata" format}
}
\value{
imported dataset
}
\description{
Imports LTM Query data and performs a set of checks that will output warnings
if file does not contain expected input, or if file contains data that may
not work with 'flfishltm' functions
}
\examples{

data(newnans)
newn <- ltm.import(newnans)
}
