% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_wt.R
\name{predict_wt}
\alias{predict_wt}
\title{Predict Weigth at Length}
\usage{
predict_wt(species_code, tl, error = 0.5)
}
\arguments{
\item{species_code}{FWC LTM species code}

\item{tl}{Total length (mm)}

\item{error}{numeric error scale expressed as a \% of estimate, defaults to 0.5}
}
\value{
tibble containing tl, prediction, upper and lower bounds
}
\description{
Predict Weigth at Length
}
\examples{
predict_wt("BLUE", 150)
predict_wt("BLUE", 150)
}
