% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_history.R
\name{species_history}
\alias{species_history}
\title{Species detection history}
\usage{
species_history(
  LTMdataset,
  exclude_species = c(),
  save = FALSE,
  filename = NA,
  fig_res = NA,
  fig_width = NA,
  fig_height = NA,
  return_object = "data"
)
}
\arguments{
\item{exclude_species}{vector of \emph{string} identifying the species names to exclude from the output figure. Should match the spelling of the species names in the figure.}

\item{save}{boolean, if TRUE figure will be saved to file}

\item{filename}{= filename for saved figure}

\item{fig_res}{= specify figure resolution}

\item{fig_width}{= specify figure width}

\item{fig_height}{= specify figure height}

\item{return_object}{\emph{string} if "data" then function returns the dataset
that is used to generate the figure, else if "ggplot" then function returns
ggplot object}

\item{ltm_dataset}{data, should be output from ltm.data.summary function}
}
\value{
either summarised data or a ggplot object. Specify the return object type with \code{return_object}
}
\description{
Create Species detection history plot
}
\examples{
data(newnans)
newn_sum <- ltm.data.summary(file=newnans)
newnans_dethist <- species.history(newn_sum,save=FALSE)
}
