% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme_ltm.R
\name{theme_ltm}
\alias{theme_ltm}
\title{Custom LTM theme for ggplot}
\usage{
theme_ltm()
}
\value{
theme for ggplot figures
}
\description{
Custom LTM theme for ggplot
}
\examples{
data(newnans)
ggplot2::ggplot(data=newnans, ggplot2::aes(x=TotalLength,y=TotalWeight)) + 
ggplot2::geom_point() +
theme_ltm()
}
