% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_June2021.R
\name{Wr}
\alias{Wr}
\title{Calculate Relative Weight}
\usage{
Wr(dataset = NA, length_col = NA, weight_col = NA, a = NA, b = NA)
}
\arguments{
\item{dataset}{dataset containing at least 2 columns, one containing a length field, and one containing weight field}

\item{length_col}{string specifying the name of the column containing length}

\item{weight_col}{string specifying the name of the column containing weights}

\item{a}{value of the alpha parameter for relative weight calculation (specific to species of interest)}

\item{b}{value of the beta parameter for relative weight calculation (specific to species of interest)}
}
\value{
returns vector of relative weights
}
\description{
Calculates the Relative Weight for a sample of fish with both lengths and weights
}
\examples{
library(dplyr)
data(newnans)
#Calculate relative weight for all LMB collected from Newnans Lake
newn_lmb <- newnans \%>\% dplyr::filter(SpeciesCode == "LMB" )
newnans_lmb_Wr <- Wr(dataset = newn_lmb, 
                 length_col = "TotalLength",
                 weight_col = "TotalWeight",
                 a = .3245,
                 b = 3.21)
}
