% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zero_count.R
\name{add_zero_count}
\alias{add_zero_count}
\title{Add zero count data for species missing from sample}
\usage{
add_zero_count(data, sample_ids, species_id, count, format = "long")
}
\arguments{
\item{data}{dataset as data.frame or tibble}

\item{sample_ids}{sample id variables}

\item{species_id}{species ID field}

\item{count}{field containing species counts}

\item{format}{defaults to "long" - data returned with 1 row per sample_id x species_id, else if not "long" returns data in "wide" format, with one row per sample_id and 1 column per unique species in the species_id.}
}
\value{
tibble
}
\description{
Add zero count data for species missing from sample
}
\examples{
fish_data <- tibble(
  site = c(1:5,3:7),
  species_name = c(rep("BLUE", 5), rep("RESU", 5)),
  ct = c(5, 2, 6, 4, 7,
         8, 6, 7, 2, 9),
  minutes = rep(10, 10))
  
fish_data \%>\% 
  add_zero_count(sample_ids = c(site,minutes), species_id = species_name, count = ct)
}
