% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_expected_columns}
\alias{check_expected_columns}
\title{HELPER: check for expected column names}
\usage{
check_expected_columns(input_table, expected_names)
}
\arguments{
\item{input_table}{}

\item{expected_names}{}
}
\description{
helper function triggers error if expected column names are missing
}
\examples{
test_data <- data.frame(col1 = 1, col2 = 2, col3 = "A")
check_expected_columns(test_data, c("col1", "col2"))
}
