% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ltm_report_helpers.R
\name{create_ltm_analysis_template}
\alias{create_ltm_analysis_template}
\title{HELPER: Create analysis R script for LTM report template}
\usage{
create_ltm_analysis_template(
  catch_data = "qryFish_standard.csv",
  age_data = NULL,
  habitat_data = NULL
)
}
\arguments{
\item{catch_data}{(string) name of a csv file containing catch data.
Data should be in format generated by 'Standard Fish Query' and stored in
the /data directory of the ltm report template.}

\item{age_data}{\href{string}{FUTURE FEATURE} name of a csv file containing age data}
}
\value{
(string) containing r script text
}
\description{
A helper function, intended to be called from within \code{\link[=create_ltm_report]{create_ltm_report()}}
}
\examples{

\dontrun{
create_ltm_analysis_template()
}


}
