% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ltm_report_helpers.R
\name{create_ltm_analysis_template}
\alias{create_ltm_analysis_template}
\title{HELPER: Create analysis R script for LTM report template}
\usage{
create_ltm_analysis_template(
  catch_data = "qryFish_standard.csv",
  age_data = NULL,
  habitat_data = "qrySite_comprehensive.csv"
)
}
\arguments{
\item{catch_data}{\emph{string} name of a csv file containing catch data.
Data should be in format generated by 'Standard Fish Query' and stored in
the /data directory of the ltm report template.}

\item{age_data}{\link{FUTURE FEATURE} \emph{string} name of a csv file containing age data}

\item{habitat_data}{\link{FUTURE FEATURE} \emph{string} name of a csv file containing habitat data.
Data should be in format generated by 'Comprehensive Site Data' query and stored in
the /data directory of the ltm report template.}
}
\value{
\emph{string} containing r script text
}
\description{
A helper function, intended to be called from within \code{\link[=create_ltm_report]{create_ltm_report()}}
}
\examples{

\dontrun{
create_ltm_analysis_template()
}


}
