% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loch_blcr_2020_age}
\alias{loch_blcr_2020_age}
\title{Lochloosa Black Crappie Fall 2020 Aged Subsample}
\format{
A dataframe with 103 obs and 5 variables
\describe{
\item{WaterBody}{Waterbody where collection was made}
\item{collected}{collection date}
\item{Species}{Species code; BLCR = Black Crappie Pomoxis nigromaculatus}
\item{tl}{total length in mm}
\item{age}{age determined from otolith reads assuming Jan 1 birthdate, in yrs}}
}
\usage{
loch_blcr_2020_age
}
\description{
Contains total length and ages for the aged subsample of Black Crappie
collected on Lochloosa Lake, FL during fall 2020 LTM trawl survyes
}
\keyword{datasets}
