% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_functions_June2021.R
\name{species.history}
\alias{species.history}
\title{Species detection history}
\usage{
species.history(
  LTMdataset,
  save = FALSE,
  filename = NA,
  fig_res = NA,
  fig_width = NA,
  fig_height = NA
)
}
\arguments{
\item{save}{boolean, if TRUE figure will be saved to file}

\item{filename}{= filename for saved figure}

\item{fig_res}{= specify figure resolution}

\item{fig_width}{= specify figure width}

\item{fig_height}{= specify figure height}

\item{ltm_dataset}{data, should be output from ltm.data.summary function}
}
\value{
returns figure
}
\description{
Create Species detection history plot
}
\examples{
data(newnans)
newn_sum <- ltm.data.summary(file=newnans)
newnans_dethist <- species.history(newn_sum,save=FALSE)
}
